Split keyvalue tests into own file and cleanup
