+++ tests/main.cpp
@@ -3,6 +3,10 @@
  * * SPDX-License-Identifier: BSD-3-Clause
  * *******************************************/
 
+#include "avx512-16bit-qsort.hpp"
+#include "avx512-32bit-qsort.hpp"
+#include "avx512-64bit-keyvaluesort.hpp"
+#include "avx512-64bit-qsort.hpp"
 #include <gtest/gtest.h>
 
 int main(int argc, char **argv)
