+++ tests/test_keyvalue.cpp
@@ -3,66 +3,17 @@
  * * SPDX-License-Identifier: BSD-3-Clause
  * *******************************************/
 
-#include "avx512-16bit-qsort.hpp"
-#include "avx512-32bit-qsort.hpp"
-#include "avx512-64bit-keyvaluesort.hpp"
-#include "avx512-64bit-qsort.hpp"
-#include "cpuinfo.h"
+#include "avx512-common-keyvaluesort.h"
 #include "rand_array.h"
 #include <gtest/gtest.h>
 #include <vector>
-
-template <typename T>
-class avx512_sort : public ::testing::Test {
-};
-TYPED_TEST_SUITE_P(avx512_sort);
-
-TYPED_TEST_P(avx512_sort, test_arrsizes)
-{
-    if (cpu_has_avx512bw()) {
-        if ((sizeof(TypeParam) == 2) && (!cpu_has_avx512_vbmi2())) {
-            GTEST_SKIP() << "Skipping this test, it requires avx512_vbmi2";
-        }
-        std::vector<int64_t> arrsizes;
-        for (int64_t ii = 0; ii < 1024; ++ii) {
-            arrsizes.push_back((TypeParam)ii);
-        }
-        std::vector<TypeParam> arr;
-        std::vector<TypeParam> sortedarr;
-        for (size_t ii = 0; ii < arrsizes.size(); ++ii) {
-            /* Random array */
-            arr = get_uniform_rand_array<TypeParam>(arrsizes[ii]);
-            sortedarr = arr;
-            /* Sort with std::sort for comparison */
-            std::sort(sortedarr.begin(), sortedarr.end());
-            avx512_qsort<TypeParam>(arr.data(), arr.size());
-            ASSERT_EQ(sortedarr, arr);
-            arr.clear();
-            sortedarr.clear();
-        }
-    }
-    else {
-        GTEST_SKIP() << "Skipping this test, it requires avx512bw";
-    }
-}
-
-REGISTER_TYPED_TEST_SUITE_P(avx512_sort, test_arrsizes);
-
-using Types = testing::Types<uint16_t,
-                             int16_t,
-                             float,
-                             double,
-                             uint32_t,
-                             int32_t,
-                             uint64_t,
-                             int64_t>;
-INSTANTIATE_TYPED_TEST_SUITE_P(TestPrefix, avx512_sort, Types);
 
 template <typename K, typename V = uint64_t>
 struct sorted_t {
     K key;
     K value;
 };
+
 template <typename K, typename V = uint64_t>
 bool compare(sorted_t<K, V> a, sorted_t<K, V> b)
 {
@@ -87,7 +38,8 @@ TYPED_TEST_P(TestKeyValueSort, KeyValueSort)
 
     for (size_t ii = 0; ii < keysizes.size(); ++ii) {
         /* Random array */
-        keys = get_uniform_rand_array_key<TypeParam>(keysizes[ii]);
+        keys =
+            get_uniform_rand_array_with_uniquevalues<TypeParam>(keysizes[ii]);
         values = get_uniform_rand_array<uint64_t>(keysizes[ii]);
         for (size_t i = 0; i < keys.size(); i++) {
             sorted_t<TypeParam, uint64_t> tmp_s;
